clean up PWM audio code
The PWM counter is based on a fixed 8MHz clock.  To have a PWM period
equal to an audio sample duration, the counter must scale up to
8MHz/samplerate. Hence:

	PWM_CLOCK = 8000000
	PWM_SCALE = PWM_CLOCK / SAMPLERATE

Then each sample pulse must have a width proportional to the actual
sample value. Our samples have a magnitude of 8 bits.  Therefore:

	pulse_width = sample * PWM_SCALE / 2^8

Because of bit truncation, the constant PWM_SCALE / 2^8 cannot be
combined. Doing the multiplication first does overflow 16 bits hence the
cast to uint32_t. And shifting the result right by 8 bits is the best way
to divide by a power of 2.

This allows for reducing in half the amount of audio interrupts incurred
by the previous code.

Also remove the unused volume code.