disable dithering when converting audio samples to 8 bits
Dithering is a technique that adds random noise to those bits that are
about to be dropped so the remaining LSB can partially represent the
signal information that would otherwise be lost. When only 8 bits remain,
this added noise represents -45 dB of the total signal which is
sufficiently important to be easily heard, especially with speech.