Let the speech generation script decide the sampling rate.
Have make.py specify the sampling rate through a define in words_def.h.
Boost it from 16000 to 18000 because there's a bit of room left in the
flash, on request from Nicolas. May need to reduce this if we ever run
out of room.
Have make.py warn if the output is too large.