don't skip last audio sample
Let's decrement the sample count only when it is consumed.

No need to safeguard against len == 0 on the first prefetch.  It was
initially there to prevent catastrophical behavior if len ended up being
negative. Now that we decrement it later, if the audio clip length is
really zero then only one byte of garbage will be played.