Try harder to interrupt ourselves when the user twiddles the controls.
Previously, we would interrupt ourselves on a major function change
(dial change) by comparing the current function name to the previous
one. However as Nicolas points out, the two voltage and 3 current dial
positions only change the scale, and so we don't react immediately. So
we'll detect those cases and interrupt, hoping as I think is the case
that in those positions, scale changes happen only by user
intervention. (This wouldn't be true for resistance or capacitance.)
Furthermore, button presses pretty much correspond to some of the
options flipping, in particular AC vs DC, so we now interrupt in those
cases too.